+++ test/test-blas.cpp
@@ -22,15 +22,15 @@ TEST(blasTest, simpleBlasCall) {
 
  
 
-   TensorVar a("a", Type(taco::UInt32, {Dimension()}), taco::dense);
-   TensorVar b("a", Type(taco::UInt32, {Dimension()}), taco::dense);
+   TensorVar a("a", Type(taco::Float32, {Dimension()}), taco::dense);
+   TensorVar b("a", Type(taco::Float32, {Dimension()}), taco::dense);
    IndexVar i("i");
 
-   Tensor<uint32_t> test_a("actuala", {3}, dense, 0);
-   Tensor<uint32_t> test_b("actualb", {3}, dense, 1);
-   Tensor<uint32_t> test_c("actualc", {3}, dense, 1);
+   Tensor<float32_t> test_a("actuala", {3}, dense, 0);
+   Tensor<float32_t> test_b("actualb", {3}, dense, 1);
+   Tensor<float32_t> test_c("actualc", {3}, dense, 1);
 
-   TensorVar workspace(Type(taco::UInt32, {1}), taco::dense, 0) ;
+   TensorVar workspace(Type(taco::Float32, {1}), taco::dense, 0) ;
 
    IndexStmt stmt = test_c(i) = test_a(i) + test_b(i);
    // IndexStmt stmt_accel = test_c(i) = test_a(i)  test_b(i);
@@ -50,19 +50,12 @@ TEST(blasTest, simpleBlasCall) {
    test_c.assemble();
    test_c.compute();
 
-   auto it = iterate<uint32_t>(test_c);
+   auto it = iterate<float32_t>(test_c);
    auto iit = it.begin();
 
    while (iit != it.end()){
       cout << "val " << iit->second << endl;
       ++iit;
    }
-
-   // test_c.evaluateAccelerated(canAccelerate);
-
-   // cout << test_c << endl;
-
-//    test_c.evaluateAccelerated(canAccelerate);
-
 }
